#95: 재치 상세 교환Cell First, Else 분리되어있던 View 합치기 및 기존 UI 코드 리팩토링
