+++ Zatch/domain/Cells/DetailCells/ShareDetailTableViewCell.swift
@@ -1,0 +1,23 @@
+//
+//  ShareDetailTableViewCell.swift
+//  Zatch
+//
+//  Created by 박지윤 on 2022/09/28.
+//
+
+import UIKit
+
+class ShareDetailTableViewCell: UITableViewCell {
+
+    override func awakeFromNib() {
+        super.awakeFromNib()
+        // Initialization code
+    }
+
+    override func setSelected(_ selected: Bool, animated: Bool) {
+        super.setSelected(selected, animated: animated)
+
+        // Configure the view for the selected state
+    }
+
+}
