+++ Zatch/domain/Views/DetailView/DetailView.swift
@@ -1,0 +1,29 @@
+//
+//  DetailView.swift
+//  Zatch
+//
+//  Created by 박지윤 on 2022/09/28.
+//
+
+import UIKit
+
+class DetailView: UIViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+
+        // Do any additional setup after loading the view.
+    }
+    
+
+    /*
+    // MARK: - Navigation
+
+    // In a storyboard-based application, you will often want to do a little preparation before navigation
+    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
+        // Get the new view controller using segue.destination.
+        // Pass the selected object to the new view controller.
+    }
+    */
+
+}
